chore: double temp
~ made consistent with iot device

+++ include/Config.h
@@ -39,7 +39,7 @@ typedef struct {
         int16_t X, Y, Z;
     } Gyro;
 
-    int16_t Temperature;
+    double Temperature;
 } MotionValues;
 
 typedef struct {
